requirements.txt
Django==3.0.7
djangorestframework==3.11.0
drf-yasg==1.17.1
psycopg2-binary==2.8.5
google-api-python-client==1.9.3
django-mailjet==0.3.1
mailjet-rest==1.3.3
django-multiselectfield==0.1.12
django-phonenumber-field==4.0.0
phonenumbers==8.12.6
Pillow==7.2.0
geocoder==1.38.1
celery==4.4.6
redis==3.5.3
gunicorn==19.6.0
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
django-storages==1.9.1
django-cors-headers==3.4.0
reportlab==3.5.46
djangorestframework-xml==2.0.0
django-phonenumber-field==4.0.0
phonenumbers==8.12.6
oauth2client==4.1.3
django-import-export==2.4.0
django-sql-explorer==2.1.2